Restructure extractors and sanitizers.
A new extract directory is available in lib to store the extractors.
Similarly, the test directory has been split into extract and sanitize
directories.

The documentURI.js sanitize file has been split into extract and
sanitize files and tests have been updated to reflect this change.
