+++ package.json
@@ -4,9 +4,9 @@
   "description": "Normalize Content Security Policy report payloads",
   "main": "index.js",
   "scripts": {
-    "test": "./node_modules/.bin/mocha -u tdd",
-    "test-cover": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- -R spec -u tdd",
-    "test-travis": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- -R spec -u tdd && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
+    "test": "./node_modules/.bin/mocha -u tdd --recursive",
+    "test-cover": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- -R spec -u tdd --recursive",
+    "test-travis": "./node_modules/.bin/istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- -R spec -u tdd  --recursive && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage",
     "lint": "jshint"
   },
   "repository": {
